test: ignoreTree getNodes
